refactor: `infra` 패키지에 있던 클래스를 `services/user/infra` 패키지 하위로 옮김
