+++ app/src/main/java/com/codesoom/assignment/services/user/UserDeleteService.java
@@ -2,7 +2,7 @@ package com.codesoom.assignment.services.user;
 
 import com.codesoom.assignment.UserNotFoundException;
 import com.codesoom.assignment.services.user.domain.User;
-import com.codesoom.assignment.infra.JpaUserRepository;
+import com.codesoom.assignment.services.user.infra.JpaUserRepository;
 import org.springframework.stereotype.Service;
 
 @Service
