+++ app/src/main/java/com/codesoom/assignment/services/user/infra/JpaUserRepository.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.infra;
+package com.codesoom.assignment.services.user.infra;
 
 import com.codesoom.assignment.services.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
